Python assessment
This is the script for a Technical Assessment I've been asked to write as part of the entry requirements for a bootcamp.

Write a program that includes: variables, if/else, functions, and some simple logic.

At the assessment please be prepared to:
-develop a simple function that will accept a few parameters and return a result
-call the simple function
-print the result
You can use any language you want at the technical assessment. If you are just learning it would be best to start with JavaScript or python. The functions we will ask for at the assessment will be very simple, for example:
-write a function that receives two parameters and returns the largest
-write a function that receives three parameters and returns the smallest